+++ src/pint/observatory/special_locations.py
@@ -53,106 +53,22 @@ class SpecialLocation(Observatory):
         self,
         name,
         aliases=None,
-        include_gps=True,
-        include_bipm=True,
-        bipm_version=bipm_default,
         overwrite=False,
     ):
-        # GPS corrections not implemented yet
-        self.include_gps = include_gps
-        self._gps_clock = None
-
-        # BIPM corrections not implemented yet
-        self.include_bipm = include_bipm
-        self.bipm_version = bipm_version
-        self._bipm_clock = None
 
         self.origin = "Built-in special location."
 
         super().__init__(name, aliases=aliases)
 
-    @property
-    def gps_fullpath(self):
-        """Returns full path to the GPS-UTC clock file.  Will first try PINT
-        data dirs, then fall back on $TEMPO2/clock."""
-        fname = "gps2utc.clk"
-        try:
-            fullpath = pint.config.runtimefile(fname)
-            return fullpath
-        except FileNotFoundError:
-            log.info(
-                "{} not found in PINT data dirs, falling back on TEMPO2/clock directory".format(
-                    fname
-                )
-            )
-            return os.path.join(os.getenv("TEMPO2"), "clock", fname)
-
-    @property
-    def bipm_fullpath(
-        self,
-    ):
-        """Returns full path to the TAI TT(BIPM) clock file.  Will first try PINT
-        data dirs, then fall back on $TEMPO2/clock."""
-        fname = "tai2tt_" + self.bipm_version.lower() + ".clk"
-        try:
-            fullpath = pint.config.runtimefile(fname)
-            return fullpath
-        except FileNotFoundError:
-            pass
-        log.info(
-            "{} not found in PINT data dirs, falling back on TEMPO2/clock directory".format(
-                fname
-            )
-        )
-        return os.path.join(os.getenv("TEMPO2"), "clock", fname)
-
-    def _load_gps_clock(self):
-        if self._gps_clock is None:
-            log.info(
-                "Observatory {0}, loading GPS clock file {1}".format(
-                    self.name, self.gps_fullpath
-                )
-            )
-            self._gps_clock = ClockFile.read(self.gps_fullpath, format="tempo2")
-
-    def _load_bipm_clock(self):
-        if self._bipm_clock is None:
-            try:
-                log.info(
-                    "Observatory {0}, loading BIPM clock file {1}".format(
-                        self.name, self.bipm_fullpath
-                    )
-                )
-                self._bipm_clock = ClockFile.read(self.bipm_fullpath, format="tempo2")
-            except:
-                raise ValueError("Can not find TT BIPM file '%s'. " % self.bipm_version)
-
     def clock_corrections(self, t, limits="warn"):
-        corr = np.zeros(t.shape) * u.s
-        if self.include_gps:
-            log.info("Applying GPS to UTC clock correction (~few nanoseconds)")
-            self._load_gps_clock()
-            corr += self._gps_clock.evaluate(t, limits=limits)
-        if self.include_bipm:
-            log.info("Applying TT(TAI) to TT(BIPM) clock correction (~27 us)")
-            self._load_bipm_clock()
-            tt2tai = 32.184 * 1e6 * u.us
-            corr += self._bipm_clock.evaluate(t, limits=limits) - tt2tai
-        return corr
+        return np.zeros(t.shape) * u.s
 
     def last_clock_correction_mjd(self):
         """Return the MJD of the last available clock correction.
 
         Returns ``np.inf`` if no clock corrections are relevant.
         """
-        t = np.inf
-        if self.include_gps:
-            self._load_gps_clock()
-            t = min(t, self._gps_clock.last_correction_mjd())
-        if self.include_bipm:
-            self._load_bipm_clock()
-            t = min(t, self._bipm_clock.last_correction_mjd())
-        return t
+        return np.inf
 
 
 class BarycenterObs(SpecialLocation):
@@ -185,13 +101,6 @@ class BarycenterObs(SpecialLocation):
             obj=self.name,
             origin="ssb",
         )
-
-    def clock_corrections(self, t, limits="warn"):
-        log.info("Special observatory location. No clock corrections applied.")
-        return np.zeros(t.shape) * u.s
-
-    def last_clock_correction_mjd(self):
-        return np.inf
 
 
 class GeocenterObs(SpecialLocation):